Lambertian to separate module
